Rename Cardano.CLI.Commands.Legacy module to Cardano.CLI.Legacy.Governance
