+++ cardano-cli/src/Cardano/CLI/EraBased/Options/Governance.hs
@@ -12,12 +12,12 @@ import           Cardano.Api
 
 import           Cardano.CLI.Environment
 import           Cardano.CLI.EraBased.Commands.Governance
-import           Cardano.CLI.EraBased.Governance
 import           Cardano.CLI.EraBased.Legacy
 import           Cardano.CLI.EraBased.Options.Common
 import           Cardano.CLI.EraBased.Options.Governance.Actions
 import           Cardano.CLI.EraBased.Options.Governance.Committee
 import           Cardano.CLI.EraBased.Vote
+import           Cardano.CLI.Legacy.Commands.Governance
 import           Cardano.CLI.Types.Common
 import           Cardano.CLI.Types.Key
 import           Cardano.CLI.Types.Legacy
