+++ cardano-cli/src/Cardano/CLI/EraBased/Run/Governance.hs
@@ -17,7 +17,7 @@ import           Cardano.Api.Shelley
 import qualified Cardano.Api.Shelley as Api
 
 import           Cardano.CLI.Commands.Governance
-import           Cardano.CLI.EraBased.Governance
+import           Cardano.CLI.Legacy.Commands.Governance
 import           Cardano.CLI.Run.Legacy.Read (fileOrPipe, readFileTx)
 import           Cardano.CLI.Types.Governance
 import qualified Cardano.CLI.Types.Governance as Cli
